+++ packages/tools/utils/src/string-boolean.spec.js
@@ -4,9 +4,14 @@ import {endsWithNewLine, isTrimmedNotEmpty} from './string-boolean';
 
 describe('String -> Boolean', function() {
 	describe('endsWithNewLine', function() {
-		it('should return true if the string ends with a newline', function() {
-			assert.deepStrictEqual(endsWithNewLine('abc'), false);
+		it('should return true if the string ends with a newline (Unix)', function() {
 			assert.deepStrictEqual(endsWithNewLine('abc\n'), true);
+		});
+		it('should return true if the string ends with a newline (Windows)', function() {
+			assert.deepStrictEqual(endsWithNewLine('abc\r\n'), true);
+		});
+		it('should return false if the string does not end with a newline', function() {
+			assert.deepStrictEqual(endsWithNewLine('abc'), false);
 		});
 	});
 	describe('isTrimmedNotEmpty', function() {