+++ packages/tools/utils/CHANGELOG.md
@@ -3,6 +3,10 @@
 - docs: converted all examples to a REPL-like format
 - dev: using single quotes rather than double quote where possible
 
+### (Any -> Any):accumcb -> (Array -> Any)
+
+- add tests to make sure `reduceFromArr` and `reduceFromObj` return a new instance of the initial value every time they get called
+
 ### (String -> String) -> (Object -> Object)
 
 - add `renameKeysWith`
@@ -30,10 +34,17 @@
 ### Object -> Boolean
 
 - add `hasSomeNullValues`
+- add tests to `endsWithNewLine` to check Windows line ending
+
+### String -> Number
+
+- add `getEndOfLineLength`
 
 ### String -> String
 
 - add `capitalize`
+- make `trimLastNewline` compatible with Windows line ending
+
 
 ## `@svizzle/utils` v0.7.1